+++ Indieplus_Pohang/View/MainFlow/DatePickerView.swift
@@ -11,13 +11,14 @@ struct DatePickerView: View {
 //    @State var selectedIndex: Int = 0
     @ObservedObject var theatermodel: TheaterManager
     @ObservedObject var moviemodel: MoviePickerViewModel
+    @ObservedObject var datemodel: DatePickerViewModel
     
     var body: some View {
         ScrollView(.horizontal, showsIndicators: false) {
             HStack{
                 Spacer()
     
-                DateView(theatermodel: theatermodel, moviemodel: moviemodel)
+                DateView(theatermodel: theatermodel, moviemodel: moviemodel, datemodel: datemodel)
                 
                 Spacer()
             }
@@ -28,74 +29,27 @@ struct DatePickerView: View {
 struct DateView: View {
     let today = Date()
     @State var selectedIndex = 0
-    @State var dateList = Set<String>()
+//    @State var dateList = Set<String>()
     @State var selectedDate = ""
     
     @StateObject var theatermodel: TheaterManager
     @StateObject var moviemodel: MoviePickerViewModel
+    @StateObject var datemodel: DatePickerViewModel
     
     @State private var count = 0
     @State private var movieTitles: [String] = []
-
-    
-    // "2023-07-01" 형태로 표시되는 날짜들만 배열에 담기
-    func dateToList(date: Date) -> String {
-        let dateFormatter = DateFormatter()
-        dateFormatter.dateFormat = "yyyy-MM-dd"
-        return dateFormatter.string(from: date)
-    }
-    
-    // 날짜 변환
-    func dateToString(date: Date) -> String {
-        let dateFormatter = DateFormatter()
-        dateFormatter.dateFormat = "d"
-        //        let day = dateFormatter.string(from: date)
-        
-        return dateFormatter.string(from: date)
-    }
-    
-    // 요일 변환
-    func dayToString(date: Date) -> String {
-        let dateFormatter = DateFormatter()
-        // 한국어로 변환
-        dateFormatter.locale = Locale(identifier: "ko_KR")
-        dateFormatter.dateFormat = "E"
-        let calendar = Calendar.current
-        
-        if calendar.isDate(date, inSameDayAs: today) {
-            return "오늘"
-        } else {
-            return dateFormatter.string(from: date)
-        }
-    }
-    
-    // 휴무일인지 판단
-    func isTodayHoliday(date: Date) -> Bool {
-        let dateFormatter = DateFormatter()
-        dateFormatter.locale = Locale(identifier: "ko_KR")
-        dateFormatter.dateFormat = "E"
-        
-        if dateFormatter.string(from: date) == "월" || dateFormatter.string(from: date) == "화" {
-            return true
-        } else {
-            DispatchQueue.main.async {
-                dateList.insert(dateToList(date: date))
-            }
-            return false
-        }
-    }
     
     var body: some View {
         ForEach(0..<10) { index in
             
             let date = Calendar.current.date(byAdding: .day, value: index, to: today)!
             
-            if !isTodayHoliday(date: date) {
+            if !datemodel.isTodayHoliday(date: date) {
                 Spacer(minLength: 10)
                 
                 Button {
                     selectedIndex = index
-                    selectedDate = dateToList(date: date)
+                    selectedDate = datemodel.dateToList(date: date)
                     moviemodel.getMovieDetail(date: selectedDate)
                     moviemodel.updateCount()
                     
@@ -112,11 +66,11 @@ struct DateView: View {
                             )
                         
                         VStack(spacing: 2){
-                            Text(dateToString(date: date))
+                            Text(datemodel.dateToString(date: date))
                                 .font(.system(size: 20, weight: .bold))
                                 .foregroundColor(index == selectedIndex ? .black : Color.main)
                             
-                            Text(dayToString(date: date))
+                            Text(datemodel.dayToString(date: date))
                                 .font(.system(size: 15, weight: .bold))
                                 .foregroundColor(index == selectedIndex ? .black : Color.main)
                         }
